Add support for variable-length strings (draft)
A variable-length string is stored as a pointer to a null-
terminated string (char *). The parser code is generated in
"ladder" structure (as a group of nested instructions).
Every "malloc" and "calloc" function has correspondent "free"
(for error cases). Return on success is done from the very
nested "if". Size checks are done for the parts of bitstream
with known size (by the moment of parsing) (not yet fully
implemented). Fixed README.md.